added a new end point for courses tree
