+++ edx_solutions_api_integration/courses/urls.py
@@ -64,6 +64,7 @@ urlpatterns = patterns(
         courses_views.CourseNavView.as_view()),
     url(r'^{0}$'.format(COURSE_ID_PATTERN), courses_views.CoursesDetail.as_view(), name='course-detail'),
     url(r'/*$^', courses_views.CoursesList.as_view()),
+    url(r'tree/*$', courses_views.CoursesTree.as_view()),
 )
 
 urlpatterns = format_suffix_patterns(urlpatterns)
